[FEAT]: created component for product card
