+++ components/index.js
@@ -1,7 +1,6 @@
 /* eslint-disable import/prefer-default-export */
 import Layout from "./global/Layout";
 import NotFound from "./global/NotFound";
-import HomePage from "./HomePage";
 import ProductDetailsPage from "./product-details/ProductDetailsPage";
 import CartItemsTable from "./cart-page/CartItemsTable";
 import CartDetailsCard from "./cart-page/CartDetailsCard";
@@ -21,11 +20,11 @@ import NoOrders from "./admin/orders/NoOrders";
 import OrdersTable from "./admin/orders/OrdersTable";
 import ProductsTable from "./admin/products/ProductsTable";
 import UserTable from "./admin/users/UserTable";
+import ProductCard from "./ProductCard";
 
 export {
   Layout,
   NotFound,
-  HomePage,
   ProductDetailsPage,
   CartItemsTable,
   CartDetailsCard,
@@ -45,4 +44,5 @@ export {
   OrdersTable,
   ProductsTable,
   UserTable,
+  ProductCard,
 };
